requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.43
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic-settings==2.1.0
httpx==0.25.2
email-validator